requirements.txt
python-dotenv==1.0.0
python-telegram-bot==13.7
pytubefix==6.7.0